requirements.txt
appdirs==1.4.3
autopep8==1.3.2
Django==1.11
django-tagulous==0.12.0
flake8==3.3.0
mccabe==0.6.1
olefile==0.44
packaging==16.8
Pillow==4.2.0
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
sorl-thumbnail==12.3
